refactor: refactor login
