+++ app/login/page.tsx
@@ -1,94 +1,15 @@
- 'use client'
+'use client'
 
-import { useState, useRef, useEffect } from 'react'
-import { useRouter } from 'next/navigation'
+import { useRef } from 'react'
 import Link from 'next/link'
-import { Building, Eye, EyeOff } from 'lucide-react'
-import toast from 'react-hot-toast'
+import { Building } from 'lucide-react'
 import AudioPlayer from '@/components/AudioPlayer'
+import LoginForm from './components/LoginForm'
+import { useBackgroundImage } from './hooks/useBackgroundImage'
 
 export default function LoginPage() {
-  const [bgImage, setBgImage] = useState<string | null>(null)
-  const [defaultBgImage, setDefaultBgImage] = useState<string | null>(null)
   const fileInputRef = useRef<HTMLInputElement | null>(null)
-
-  useEffect(() => {
-    // Load persisted background from localStorage if present (has highest priority)
-    try {
-      const stored = localStorage.getItem('loginBackground')
-      if (stored) {
-        setBgImage(stored)
-        return // User custom image takes priority
-      }
-    } catch (e) {
-      // ignore
-    }
-
-    // Try to load default background image from assets
-    const checkDefaultImage = () => {
-      // Try different image formats in order
-      const formats = ['jpg', 'jpeg', 'png', 'webp']
-      let formatIndex = 0
-      
-      const tryNextFormat = () => {
-        if (formatIndex >= formats.length) return
-        
-        const img = new Image()
-        const format = formats[formatIndex]
-        img.src = `/assets/images/backgrounds/login.${format}`
-        
-        img.onload = () => {
-          setDefaultBgImage(img.src)
-        }
-        
-        img.onerror = () => {
-          formatIndex++
-          tryNextFormat()
-        }
-      }
-      
-      tryNextFormat()
-    }
-    
-    checkDefaultImage()
-  }, [])
-  const [email, setEmail] = useState('')
-  const [password, setPassword] = useState('')
-  const [showPassword, setShowPassword] = useState(false)
-  const [loading, setLoading] = useState(false)
-  const router = useRouter()
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault()
-    setLoading(true)
-
-    try {
-      const response = await fetch('/api/auth/login', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ email, password }),
-      })
-
-      const data = await response.json()
-
-      if (response.ok) {
-        toast.success('Đăng nhập thành công!')
-        console.log('Login successful, redirecting to dashboard...')
-        // Use window.location for more reliable redirect
-        setTimeout(() => {
-          window.location.href = '/dashboard'
-        }, 1000)
-      } else {
-        toast.error(data.message || 'Đăng nhập thất bại!')
-      }
-    } catch (error) {
-      toast.error('Có lỗi xảy ra, vui lòng thử lại!')
-    } finally {
-      setLoading(false)
-    }
-  }
+  const { backgroundStyle, hasAnyImage, setBgImage } = useBackgroundImage()
 
   const triggerFileSelect = () => {
     fileInputRef.current?.click()
@@ -113,18 +34,13 @@ export default function LoginPage() {
   return (
     <div
       className="min-h-screen flex items-center justify-center py-12 px-4 sm:px-6 lg:px-8 relative"
-      style={{
-        backgroundImage: bgImage || defaultBgImage ? `url(${bgImage || defaultBgImage})` : undefined,
-        backgroundSize: 'cover',
-        backgroundPosition: 'center',
-        backgroundRepeat: 'no-repeat'
-      }}
+      style={backgroundStyle}
     >
-      {/* Background overlay when no custom image */}
-      {!bgImage && !defaultBgImage && (
+      {/* Background overlay */}
+      {!hasAnyImage && (
         <div className="absolute inset-0 bg-gradient-to-br from-yellow-2 via-white to-yellow-2"></div>
       )}
-      {(bgImage || defaultBgImage) && (
+      {hasAnyImage && (
         <div className="absolute inset-0 bg-gray-900/40 backdrop-blur-sm"></div>
       )}
 
@@ -157,69 +73,7 @@ export default function LoginPage() {
                 </Link>
               </p>
             </div>
-            <form className="mt-8 space-y-6" onSubmit={handleSubmit}>
-              <div className="space-y-4">
-                <div>
-                  <label htmlFor="email" className="block text-sm font-semibold text-gray-700 mb-2">
-                    Email
-                  </label>
-                  <input
-                    id="email"
-                    name="email"
-                    type="email"
-                    autoComplete="email"
-                    required
-                    className="w-full px-4 py-3 border border-gray-200 rounded-[8px] bg-gray-50 text-gray-900 placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-navy-1 focus:border-transparent transition-all duration-200"
-                    placeholder="Nhập địa chỉ email"
-                    value={email}
-                    onChange={(e) => setEmail(e.target.value)}
-                  />
-                </div>
-                <div className="relative">
-                  <label htmlFor="password" className="block text-sm font-semibold text-gray-700 mb-2">
-                    Mật khẩu
-                  </label>
-                  <input
-                    id="password"
-                    name="password"
-                    type={showPassword ? 'text' : 'password'}
-                    autoComplete="current-password"
-                    required
-                    className="w-full px-4 py-3 pr-12 border border-gray-200 rounded-[8px] bg-gray-50 text-gray-900 placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-navy-1 focus:border-transparent transition-all duration-200"
-                    placeholder="Nhập mật khẩu"
-                    value={password}
-                    onChange={(e) => setPassword(e.target.value)}
-                  />
-                  <button
-                    type="button"
-                    className="absolute right-3 bottom-3 flex items-center text-gray-400 hover:text-gray-600 transition-colors"
-                    onClick={() => setShowPassword(!showPassword)}
-                  >
-                    {showPassword ? (
-                      <EyeOff className="h-5 w-5" />
-                    ) : (
-                      <Eye className="h-5 w-5" />
-                    )}
-                  </button>
-                </div>
-              </div>
-
-              <div>
-                <button
-                  type="submit"
-                  disabled={loading}
-                  className="w-full flex justify-center py-3 px-4 border border-transparent text-base font-semibold rounded-[8px] text-white bg-gradient-to-r from-navy-1 to-navy-2 hover:shadow-drop-lg focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-navy-1 disabled:opacity-50 disabled:cursor-not-allowed transition-all duration-200 transform hover:-translate-y-0.5"
-                >
-                  {loading ? 'Đang đăng nhập...' : 'Đăng nhập'}
-                </button>
-              </div>
-
-              <div className="text-center">
-                <Link href="/" className="text-sm text-navy-1 hover:text-navy-2 font-medium transition-colors">
-                  ← Quay lại trang chủ
-                </Link>
-              </div>
-            </form>
+            <LoginForm />
           </div>
         </div>
       </div>
